steal_finder/requirements.txt
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
googlemaps==4.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
iniconfig==2.0.0
jiter==0.7.0
packaging==24.1
playwright==1.48.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tokenizers==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
